Fix projectile project cache loading.
After the previous change, this made it so we wouldn't need to use
projectile-global-mode, I uncovered an issue: projectile-mode must be
loaded at least once for project caches to work correctly. This
changelist turns on projectile-mode as a part of load-project-management
to do project cache initialization (and possibly some other
initialization), and then turns it off again.