Create named-ansi-term function.
This function is a wrapper for ansi-term. It interactively asks for the
same inputs for which ansi-term would ask, but it customizes the buffer
name so that is has the form "*ansi-<current directory>*".

This function is now assigned to C-c s a in place of ansi-term, and
ansi-term-in-directory now uses this instead of ansi-term.