Fix C/C++ styling.
There was a typo in c-default-style. Also, apparently it isn't enough to
just set the default style, so explicitly set the C style to "linux",
which more closely matches what I need. I use 4 as the offset instead of
the linux style's 8, so that gets set after setting the style.