Update close confirmation code.
save-buffers-kill-terminal, which is bound to C-x C-c, checks a variable
called confirm-kill-emacs (nil by default). This variable can be set to
a close confirmation function. This commit adds an interactive function
called toggle-close-confirm, which toggles the value of
confirm-kill-emacs between nil and 'yes-or-no-p. This changelist also
removes confirm-emacs-close, which is an old function that I wrote that
I was no longer using.

toggle-close-confirm is not currently bound to any key combination.