Make running VNC optional.
By default the image will be run interactively. The entrypoint.sh script has
been renamed to run_vnc.sh, which can be invoked as an alternative comment to
the regular entrypoint. The script still sleeps forever in the background after
running everything, so if it is run interactively it will have to be run in the
background if other commands need to be run after.