Only attempt to load vterm when cmake is installed.
cmake, along with other build tools, is required for building libvterm. This
change checks to make sure that cmake is installed before attempting to
load vterm and possibly build libvterm.

Of course, this is just an simplification of what is required to load vterm
successfully, but is good enough for most of my use cases.